requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
tensorflow>=2.13.0

# Time Series
statsmodels>=0.14.0
prophet>=1.1.5
pmdarima>=2.0.0

# Visualization
streamlit>=1.28.0
plotly>=5.17.0
folium>=0.15.0
seaborn>=0.13.0
matplotlib>=3.8.0
kaleido>=0.2.1

# Model Interpretation
shap>=0.43.0

# Utilities
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Additional ML Libraries
imbalanced-learn>=0.11.0
optuna>=3.4.0
